Move chart drill-down page to index
